feat(ai): Modularize DreamService into standalone singletons (#10028)
