+++ ai/daemons/DreamService.mjs
@@ -16,6 +16,8 @@ import GapInferenceEngine      from './services/GapInferenceEngine.mjs';
 import GoldenPathSynthesizer   from './services/GoldenPathSynthesizer.mjs';
 import GraphMaintenanceService from './services/GraphMaintenanceService.mjs';
 import IssueIngestor           from './services/IssueIngestor.mjs';
+import SemanticGraphExtractor  from './services/SemanticGraphExtractor.mjs';
+import TopologyInferenceEngine from './services/TopologyInferenceEngine.mjs';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname  = path.dirname(__filename);
@@ -185,12 +187,12 @@ class DreamService extends Base {
                 logger.info(`[DreamService]   -> Payload size (chars): ${session.document.length}`);
                 
                 const startTime = Date.now();
-                const success = await this.executeTriVectorExtraction(session);
+                const success = await SemanticGraphExtractor.executeTriVectorExtraction(session);
                 const triVectorTime = ((Date.now() - startTime) / 1000).toFixed(1);
                 logger.info(`[DreamService]   -> Tri-Vector Synthesis took: ${triVectorTime}s`);
 
                 const topoStart = Date.now();
-                await this.extractTopology(session.document, session.meta.sessionId);
+                await TopologyInferenceEngine.extractTopology(session.document, session.meta.sessionId);
                 const topoTime = ((Date.now() - topoStart) / 1000).toFixed(1);
                 logger.info(`[DreamService]   -> Topological Conflicts took: ${topoTime}s`);
                 
@@ -222,295 +224,6 @@ class DreamService extends Base {
             logger.error('[DreamService] Failed to process undigested sessions:', error);
         } finally {
             this.isProcessing = false;
-        }
-    }
-
-    /**
-     * Executes the Tri-Vector Synthesis (Semantic Graph, Open Deltas, Roadmap Strategy)
-     * from the session memory log via JSON schema extraction.
-     * @param {Object} session Wrapped session object containing id, document, and meta
-     * @returns {Promise<Object|null>} The extracted payload, or null on failure
-     */
-    async executeTriVectorExtraction(session) {
-        logger.info(`[DreamService] Extracting Tri-Vector Synthesis for session ID: ${session.meta.sessionId}`);
-
-        const systemInstruction = `You are the Neo.mjs REM (Rapid Eye Movement) Sleep digestion agent.
-Your task is to analyze the following episodic development session history and extract three vital vectors of intelligence into a strict A2A 2026 JSON object:
-
-1. **Semantic Graph:** Core concepts, framework components, and their relationships.
-2. **Feature Namespace:** What primary class or namespace were we working on?
-3. **Human Readable Summary:** A single sentence summary of the turn/session.
-
-Enforce this STRICT JSON schema:
-{
-  "a2a_version": "1.0",
-  "agent_id": "Antigravity_Primary",
-  "session_artifact": {
-    "feature_namespace": "String (e.g. Neo.dashboard.Main, or null)",
-    "human_readable_summary": "String (1 sentence high-level summary of the session or turn)",
-    "roadmap_impact": "String (Proposal for a long-term strategy pivot) or null",
-    "graph": {
-      "nodes": [
-        {
-          "id": "Type:Name",
-          "type": "String (MUST BE EXACTLY ONE OF: SESSION, MEMORY, ARTIFACT_PLAN, ARTIFACT_TASK, ISSUE, STRATEGY, SYSTEM_ANCHOR, CONCEPT, CLASS, METHOD, FILE, GUIDE, BLOG, TEST)",
-          "name": "String",
-          "description": "String",
-          "logical_layer": "String (e.g. UI, State, Network, Build, Docs, Core, Unknown)",
-          "stability": "String (EXPERIMENTAL, STABLE, DEPRECATED, UNKNOWN)",
-          "gravity_well": "Boolean (Is this a long-term strategic anchor from roadmap/boardroom?)",
-          "strategic_weight": 0.9,
-          "confidence": 0.9,
-          "tags": ["Array", "of", "Strings"]
-        }
-      ],
-      "edges": [
-        {
-          "source": "String (must match a node id, or 'frontier')",
-          "target": "String (must match a node id, or 'frontier')",
-          "relationship": "String (MUST BE EXACTLY ONE OF: IMPLEMENTS, EXTENDS, DEPENDS_ON, BLOCKS, BLOCKED_BY, RELATES_TO, RESOLVES, CAUSES_ISSUE)",
-          "weight": 1.0,
-          "justification": "String (Brief reason for this edge's algorithmic relevance)"
-        }
-      ]
-    }
-  }
-}
-
-DO NOT output markdown, \`\`\`json blocks, or any other explanations. Provide purely the JSON object.`;
-
-        try {
-            const provider = Neo.create(OpenAiCompatible, {
-                modelName: aiConfig.openAiCompatible.model,
-                host: aiConfig.openAiCompatible.host
-            });
-
-            // Format boundaries securely
-            const messages = [
-                { role: 'system', content: systemInstruction },
-                { role: 'user', content: `--- Session Episodic Memory ---\n${session.document}` }
-            ];
-
-            let maxRetries = 3;
-            let attempt = 0;
-            let payload = null;
-            let result = null;
-
-            while (attempt < maxRetries && !payload) {
-                attempt++;
-                
-                // Call standard generation method explicitly without format enforcement
-                result = await provider.generate(messages);
-
-                // Extract using robust Json parser to catch malformed boundaries
-                payload = Json.extract(result.content);
-
-                // Validation check
-                if (!payload || !payload.session_artifact || !payload.session_artifact.graph || !payload.session_artifact.graph.nodes || !payload.session_artifact.graph.edges) {
-                    logger.warn(`[DreamService] Attempt ${attempt}: Failed to validate extracted Tri-Vector A2A payload for session: ${session.meta.sessionId}`);
-                    
-                    if (attempt < maxRetries) {
-                        logger.warn(`[DreamService] Attempt ${attempt}: Injecting autonomous JSON repair feedback loop.`);
-                        messages.push({ role: 'assistant', content: result.content });
-                        messages.push({ 
-                            role: 'user', 
-                            content: `Your previous response failed internal schema validation. You are either missing required keys (e.g., session_artifact, graph.nodes, graph.edges) or you provided malformed JSON. Please correct your output and provide ONLY the exact JSON shape requested in the instructions.`
-                        });
-                        payload = null; // Ensure loop continues
-                    } else {
-                        logger.warn(`[DreamService] --- FINAL EXHAUSTED RAW LLM DUMP ---\n${result.content}\n-----------------------------`);
-                    }
-                }
-            }
-            
-            if (!payload) {
-                return null;
-            }
-
-            logger.debug(`[DreamService] Successfully extracted Tri-Vector A2A schema for session ${session.meta.sessionId} after ${attempt} attempts.`);
-
-            const artifact = payload.session_artifact;
-
-            // --- VECTOR 1: SEMANTIC GRAPH ---
-            // Ensure frontier exists, if not, stub it so we can link to it
-            if (!GraphService.db.nodes.has('frontier')) {
-                GraphService.upsertNode({
-                    id: 'frontier',
-                    type: 'SYSTEM_ANCHOR',
-                    name: 'Active Context Frontier',
-                    description: 'The actively tracked development front for the current project scope.',
-                    semanticVectorId: null
-                });
-            }
-
-            const VALID_TYPES = ['SESSION', 'MEMORY', 'ARTIFACT_PLAN', 'ARTIFACT_TASK', 'ISSUE', 'STRATEGY', 'SYSTEM_ANCHOR', 'CONCEPT', 'CLASS', 'METHOD', 'FILE', 'GUIDE', 'BLOG', 'TEST'];
-
-            // Bridge to GraphService (SQLite)
-            for (const node of artifact.graph.nodes) {
-                if (node.id === 'frontier') continue;
-                
-                let nodeType = node.type && VALID_TYPES.includes(node.type.toUpperCase()) ? node.type.toUpperCase() : 'CONCEPT';
-                let nodeId = node.id;
-                
-                // Enforce Neo native Graph ID specification (Type:Name) if hallucinated
-                if (!nodeId.includes(':')) {
-                    const cleanName = (node.name || nodeId).replace(/[^a-zA-Z0-9_\-\.]/g, '_');
-                    nodeId = `${nodeType}:${cleanName}`;
-                }
-
-                GraphService.upsertNode({
-                    id: nodeId,
-                    type: nodeType,
-                    name: node.name || 'Unknown',
-                    description: node.description || '',
-                    semanticVectorId: session.id,
-                    properties: {
-                        logical_layer: node.logical_layer || 'Unknown',
-                        stability: node.stability || 'UNKNOWN',
-                        gravity_well: node.gravity_well === true,
-                        strategic_weight: typeof node.strategic_weight === 'number' ? node.strategic_weight : (node.gravity_well ? 1.0 : 0.1),
-                        confidence: typeof node.confidence === 'number' ? node.confidence : 0.5,
-                        tags: Array.isArray(node.tags) ? node.tags : [],
-                        context_source: session.meta.sessionId
-                    }
-                });
-                
-                // Update the payload graph node id so edges bind correctly
-                node._resolvedId = nodeId; 
-            }
-
-            const validNodeRefs = new Set([...artifact.graph.nodes.map(n => n.id), ...artifact.graph.nodes.map(n => n._resolvedId), 'frontier']);
-
-            for (const edge of artifact.graph.edges) {
-                // Map the original edge source/target to the resolved Node IDs
-                let resolvedSource = edge.source;
-                let resolvedTarget = edge.target;
-                
-                const sourceNode = artifact.graph.nodes.find(n => n.id === edge.source);
-                if (sourceNode && sourceNode._resolvedId) resolvedSource = sourceNode._resolvedId;
-                
-                const targetNode = artifact.graph.nodes.find(n => n.id === edge.target);
-                if (targetNode && targetNode._resolvedId) resolvedTarget = targetNode._resolvedId;
-
-                if (!validNodeRefs.has(resolvedSource) || !validNodeRefs.has(resolvedTarget)) {
-                    logger.warn(`[DreamService] Culling hallucinated edge from ${resolvedSource} to ${resolvedTarget}`);
-                    continue; // Skip trying to link non-existent graph nodes
-                }
-
-                GraphService.linkNodes(
-                    resolvedSource,
-                    resolvedTarget,
-                    edge.relationship || 'RELATES_TO',
-                    edge.weight !== undefined ? parseFloat(edge.weight) : 1.0,
-                    {
-                        justification: edge.justification || '',
-                        context_source: session.meta.sessionId
-                    }
-                );
-            }
-
-            logger.info(`[DreamService] Graph entities committed to Neocortex for session ${session.meta.sessionId}.`);
-
-            // --- VECTOR 2: STRATEGIC ROADMAP PIVOTS ---
-            if (artifact.roadmap_impact && typeof artifact.roadmap_impact === 'string' && artifact.roadmap_impact.toLowerCase() !== 'null') {
-                const auditLog = path.join('/tmp', 'roadmap_audits.log');
-                const strategyEntry = `[${new Date().toISOString()}] Session ${session.meta.sessionId}:\n${artifact.roadmap_impact}\n\n`;
-                await fs.promises.appendFile(auditLog, strategyEntry, 'utf8');
-                logger.info(`[DreamService] Extracted Strategy impact to roadmap_audits.log`);
-            }
-
-            return payload;
-
-        } catch (error) {
-            if (error.message && error.message.includes('fetch failed')) {
-                logger.debug(`[DreamService] Skipping extraction (API provider offline).`);
-            } else {
-                logger.error('[DreamService] Error during graph extraction run:', error);
-            }
-            return null;
-        }
-    }
-
-    /**
-     * Dedicated inference pass to scan episodic memory explicitly for topological conflicts
-     * (e.g. tracking when an OPEN issue is superseded or rendered obsolete by recent session decisions).
-     * @param {String} contextText The raw session episodic document.
-     * @param {String} sessionId The ID of the session being processed.
-     */
-    async extractTopology(contextText, sessionId) {
-        logger.info(`[DreamService] Extracting Topological Conflicts for session ID: ${sessionId}`);
-
-        const prompt = `
-You are the Neo.mjs REM Sandman. Analyze the following session history for strict topological conflicts.
-A topological conflict occurs primarily when the user and agent realize an OPEN GitHub ticket/issue has been rendered obsolete, superseded, or is a duplicate.
-
-Enforce this STRICT JSON schema:
-{
-  "conflicts": [
-    {
-      "issueId": "String (e.g. issue-1234)",
-      "type": "String (SUPERSEDES, OBSOLETES, DUPLICATE)",
-      "description": "String (Why is there a conflict?)"
-    }
-  ]
-}
-
-DO NOT output markdown, \`\`\`json blocks, or any other explanations. Provide purely the JSON object. If there are no conflicts, output {"conflicts": []}.
-
---- Session Episodic Memory ---
-${contextText}
-`;
-        try {
-            const provider = Neo.create(OpenAiCompatible, {
-                modelName: aiConfig.openAiCompatible.model,
-                host: aiConfig.openAiCompatible.host
-            });
-
-            const result = await provider.generate(prompt);
-
-            const payload = Json.extract(result.content);
-            if (!payload || !Array.isArray(payload.conflicts) || payload.conflicts.length === 0) {
-                return;
-            }
-
-            // Write to sandman_handoff.md
-            const handoffFile = aiConfig.handoffFilePath;
-            const tmpFile = `${handoffFile}.tmp`;
-
-            let handoffContent = '';
-            try {
-                handoffContent = await fs.promises.readFile(handoffFile, 'utf8');
-            } catch (e) {
-                handoffContent = '# Sandman Handoff Alerts\n\nThis file tracks topological conflict alerts generated during overnight REM sleep cycles. Agents MUST reconcile these conflicts structurally upon startup.\n\n## Active Conflicts\n\n';
-            }
-
-            let newAlerts = false;
-            for (const conflict of payload.conflicts) {
-                const entry = `- **[${conflict.type}]** \`${conflict.issueId}\`: ${conflict.description} (Source Session: ${sessionId})\n`;
-                const anyConflictIdentifier = `\`${conflict.issueId}\`:`;
-                if (!handoffContent.includes(anyConflictIdentifier)) {
-                    let insertIndex = handoffContent.indexOf('## Computed Golden Path');
-                    if (insertIndex !== -1) {
-                        handoffContent = handoffContent.substring(0, insertIndex) + entry + '\n\n' + handoffContent.substring(insertIndex);
-                    } else {
-                        handoffContent += entry;
-                    }
-                    newAlerts = true;
-                }
-            }
-
-            if (newAlerts) {
-                await fs.promises.writeFile(tmpFile, handoffContent, 'utf8');
-                await fs.promises.rename(tmpFile, handoffFile);
-                logger.info(`[DreamService] Registered new topological conflicts to sandman_handoff.md for session ${sessionId}.`);
-            }
-
-        } catch (error) {
-            if (error.message && error.message.includes('fetch failed')) {
-                logger.debug('[DreamService] Skipping topology extraction (API provider offline).');
-            } else {
-                logger.error('[DreamService] Error during topology extraction:', error);
-            }
         }
     }
